fix: restructured folders
