+++ src/projects/pretrain/config.cfg
@@ -1,6 +1,6 @@
 [training]
 batch_size=64
-num_workers_for_dataloader= 1
+num_workers_for_dataloader= 8
 
 [training.trainer]
 accelerator = "auto"
@@ -13,7 +13,7 @@ min_epochs = null
 max_steps = -1
 min_steps = null
 limit_train_batches = null
-limit_val_batches = null
+limit_val_batches = 100
 limit_test_batches = null
 limit_predict_batches = null
 overfit_batches = 0.0
