+++ src/snp_transformer/model/task_modules/__init__.py
@@ -1,0 +1,2 @@
+from .encoder_for_masked_lm import EncoderForMaskedLM, create_encoder_for_masked_lm
+from .trainable_modules import MaskingTargets, TrainableModule
