+++ tests/test_train.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
 import pytest
+
 from snp_transformer.train import train
